+++ lib/rack/service_api_versioning/invalid_base_url_error.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+# Exception wrapper class for invalid return data due to bad SBU. This is not
+# nested within the `EncodedApiVersionData` class so as not to unnecessarily
+# leak extraneous implementation details.
+class InvalidBaseUrlError < RuntimeError
+  def initialize(base_url, original_error)
+    @original_error = original_error
+    super "Invalidly formatted base URL: #{base_url}"
+  end
+
+  attr_reader :original_error
+end # class InvalidBaseUrlError
